Addet Paintbrush and clicked startbutton texture
I will add more features of paintbrush soon. For now, youre able to
draw with a pen and brush, making boxes and circles and erase parts of the picture with the rubber tool. You can also pick a color with the color picking tool. And you can
do File->New..
I hope you Accept this.